Refactor code to improve structure and remove unused imports as well as documentation
